Properly support large gltf models. Getting cooked by a weird texture bug. Cleanup and new gltf files.

+++ src/renderer/vk_types.h
@@ -46,7 +46,6 @@ struct Vertex
     glm::vec4 color{1.0f};
     glm::vec2 uv{0,0};
     glm::uint32_t materialIndex{0};
-
 };
 
 struct Mesh
@@ -61,8 +60,8 @@ struct Material
 {
     glm::vec4 colorFactor{1.0f};
     glm::vec4 metalRoughFactors{0.0f, 1.0f, 0.0f, 0.0f}; // x: metallic, y: roughness
-    glm::ivec4 textureImageIndices{-1};   // x: color image, y: metallic image, z: pad, w: pad
-    glm::ivec4 textureSamplerIndices{-1}; // x: color sampler, y: metallic sampler, z: pad, w: pad
+    glm::vec4 textureImageIndices{-1};   // x: color image, y: metallic image, z: pad, w: pad
+    glm::vec4 textureSamplerIndices{-1}; // x: color sampler, y: metallic sampler, z: pad, w: pad
     glm::vec4 alphaCutoff{1.0f, 0.0f, 0.0f, 0.0f}; // x: alpha cutoff, y: alpha mode, z: padding, w: padding
 };
